refactor: split control plane acp mapping
